+++ cognite/client/data_classes/annotation_types/primitives.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, Literal, cast
 
 from cognite.client.data_classes._base import CogniteResource
-from cognite.client.utils._auxiliary import local_import
+from cognite.client.utils._importing import local_import
 from cognite.client.utils._text import convert_dict_to_case
 
 if TYPE_CHECKING: