+++ cognite/client/data_classes/data_modeling/query.py
@@ -16,7 +16,7 @@ from cognite.client.data_classes.data_modeling.instances import (
     PropertyValue,
 )
 from cognite.client.data_classes.filters import Filter
-from cognite.client.utils._auxiliary import local_import
+from cognite.client.utils._importing import local_import
 
 
 @dataclass