+++ cognite/client/utils/_auxiliary.py
@@ -1,13 +1,11 @@
 from __future__ import annotations
 
 import functools
-import importlib
 import math
 import numbers
 import platform
 import warnings
 from decimal import Decimal
-from types import ModuleType
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -20,7 +18,6 @@ from typing import (
 )
 from urllib.parse import quote
 
-from cognite.client.exceptions import CogniteImportError
 from cognite.client.utils._text import (
     convert_all_keys_to_camel_case,
     convert_all_keys_to_snake_case,
@@ -36,6 +33,10 @@ if TYPE_CHECKING:
 
 T = TypeVar("T")
 THashable = TypeVar("THashable", bound=Hashable)
+
+
+def no_op(x: T) -> T:
+    return x
 
 
 def is_unlimited(limit: float | int | None) -> bool:
@@ -110,60 +111,8 @@ def json_dump_default(x: Any) -> Any:
     raise TypeError(f"Object {x} of type {x.__class__} can't be serialized by the JSON encoder")
 
 
-def assert_type(var: Any, var_name: str, types: list[type], allow_none: bool = False) -> None:
-    if var is None:
-        if not allow_none:
-            raise TypeError(f"{var_name} cannot be None")
-    elif not isinstance(var, tuple(types)):
-        raise TypeError(f"{var_name!r} must be one of types {types}, not {type(var)}")
-
-
 def interpolate_and_url_encode(path: str, *args: Any) -> str:
     return path.format(*[quote(str(arg), safe="") for arg in args])
-
-
-@overload
-def local_import(m1: str, /) -> ModuleType:
-    ...
-
-
-@overload
-def local_import(m1: str, m2: str, /) -> tuple[ModuleType, ModuleType]:
-    ...
-
-
-@overload
-def local_import(m1: str, m2: str, m3: str, /) -> tuple[ModuleType, ModuleType, ModuleType]:
-    ...
-
-
-@overload
-def local_import(m1: str, m2: str, m3: str, m4: str, /) -> tuple[ModuleType, ModuleType, ModuleType, ModuleType]:
-    ...
-
-
-def local_import(*module: str) -> ModuleType | tuple[ModuleType, ...]:
-    assert_type(module, "module", [tuple])
-    if len(module) == 1:
-        name = module[0]
-        try:
-            return importlib.import_module(name)
-        except ImportError as e:
-            raise CogniteImportError(name.split(".")[0]) from e
-
-    modules = []
-    for name in module:
-        try:
-            modules.append(importlib.import_module(name))
-        except ImportError as e:
-            raise CogniteImportError(name.split(".")[0]) from e
-    return tuple(modules)
-
-
-def import_legacy_protobuf() -> bool:
-    from google.protobuf import __version__ as pb_version
-
-    return 4 > int(pb_version.split(".")[0])
 
 
 def get_current_sdk_version() -> str: