+++ cognite/client/utils/_concurrency.py
@@ -2,14 +2,10 @@ from __future__ import annotations
 
 import functools
 from collections import UserList
-from concurrent.futures import CancelledError, ThreadPoolExecutor
-from copy import copy
+from concurrent.futures import ThreadPoolExecutor
 from typing import (
-    TYPE_CHECKING,
     Any,
     Callable,
-    Iterable,
-    Iterator,
     Literal,
     Protocol,
     Sequence,
@@ -17,26 +13,7 @@ from typing import (
 )
 
 from cognite.client.exceptions import CogniteAPIError, CogniteDuplicatedError, CogniteNotFoundError
-
-if TYPE_CHECKING:
-    from concurrent.futures import Future
-    from types import TracebackType
-
-_T = TypeVar("_T")
-
-
-def import_as_completed() -> Callable[[Iterable[Future[_T]]], Iterator[Future[_T]]]:
-    from cognite.client._constants import _RUNNING_IN_BROWSER
-
-    if not _RUNNING_IN_BROWSER:
-        from concurrent.futures import as_completed
-    else:
-        from copy import copy
-
-        def as_completed(fs: Iterable[Future[_T]], timeout: float | None = None) -> Iterator[Future[_T]]:
-            return iter(copy(fs))
-
-    return as_completed
+from cognite.client.utils._auxiliary import no_op
 
 
 class TasksSummary:
@@ -50,7 +27,7 @@ class TasksSummary:
         self.exceptions = exceptions
 
     def joined_results(self, unwrap_fn: Callable | None = None) -> list:
-        unwrap_fn = unwrap_fn or (lambda x: x)
+        unwrap_fn = unwrap_fn or no_op
         joined_results: list = []
         for result in self.results:
             unwrapped = unwrap_fn(result)
@@ -68,7 +45,7 @@ class TasksSummary:
     ) -> None:
         if not self.exceptions:
             return None
-        task_unwrap_fn = (lambda x: x) if task_unwrap_fn is None else task_unwrap_fn
+        task_unwrap_fn = no_op if task_unwrap_fn is None else task_unwrap_fn
         if task_list_element_unwrap_fn is not None:
             successful = []
             for t in self.successful_tasks:
@@ -89,6 +66,7 @@ class TasksSummary:
         )
 
     def raise_first_encountered_exception(self) -> None:
+        # TODO: Change to 'raise_compound_exception_if_failed_tasks' in next major version?
         if self.exceptions:
             raise self.exceptions[0]
 
@@ -162,17 +140,11 @@ class SyncFuture(TaskFuture[T_Result]):
     def __init__(self, fn: Callable[..., T_Result], *args: Any, **kwargs: Any) -> None:
         self._task = functools.partial(fn, *args, **kwargs)
         self._result: T_Result | None = None
-        self._is_cancelled = False
 
     def result(self) -> T_Result:
-        if self._is_cancelled:
-            raise CancelledError
         if self._result is None:
             self._result = self._task()
         return self._result
-
-    def cancel(self) -> None:
-        self._is_cancelled = True
 
 
 class MainThreadExecutor(TaskExecutor):
@@ -193,24 +165,6 @@ class MainThreadExecutor(TaskExecutor):
 
     def submit(self, fn: Callable[..., T_Result], *args: Any, **kwargs: Any) -> SyncFuture:
         return SyncFuture(fn, *args, **kwargs)
-
-    def shutdown(self, wait: bool = False) -> None:
-        return None
-
-    @staticmethod
-    def as_completed(it: Iterable[SyncFuture]) -> Iterator[SyncFuture]:
-        return iter(copy(it))
-
-    def __enter__(self) -> MainThreadExecutor:
-        return self
-
-    def __exit__(
-        self,
-        type: type[BaseException] | None,
-        value: BaseException | None,
-        traceback: TracebackType | None,
-    ) -> None:
-        self.shutdown()
 
 
 _THREAD_POOL_EXECUTOR_SINGLETON: ThreadPoolExecutor