+++ cognite/client/utils/_version_checker.py
@@ -6,8 +6,6 @@ import re
 
 import requests
 
-from cognite.client.config import global_config
-
 
 def check_if_version_exists(package_name: str, version: str) -> bool:
     versions = get_all_versions(package_name=package_name)
@@ -25,6 +23,8 @@ def get_newest_version_in_major_release(package_name: str, version: str) -> str:
 
 
 def get_all_versions(package_name: str) -> list[str]:
+    from cognite.client.config import global_config
+
     verify_ssl = not global_config.disable_ssl
     res = requests.get(f"https://pypi.python.org/simple/{package_name}/#history", verify=verify_ssl, timeout=5)
     return re.findall(r"cognite-sdk-(\d+\.\d+.[\dabrc]+)", res.content.decode())