+++ tests/tests_unit/test_api/test_geospatial.py
@@ -4,7 +4,7 @@ import uuid
 import pytest
 
 from cognite.client.data_classes.geospatial import Feature, FeatureList, FeatureType
-from cognite.client.utils._auxiliary import local_import
+from cognite.client.utils._importing import local_import
 
 
 @pytest.fixture()